Found multi threading bug
When executing server and client in the same process as different
threads, we seem to be stuck in a spin-lock. This does seem to not (or
less often) happen for latency tests. Maybe caused by the regular call
to chrono?

The problem disapears when calling server and client as different
processes.